Refactored Annotations\Reflection tests.
